getOpening/Force -> float. - version of 2014-05-27 -

+++ wsg_50/include/wsg_50/functions.h
@@ -69,8 +69,8 @@ int setGraspingForceLimit( float force );
 
 const char * systemState( void );
 int graspingState( void );
-int getOpening( void );
-int getForce( void );
+float getOpening( void );
+float getForce( void );
 int getAcceleration( void );
 int getGraspingForceLimit( void );
